Refactor: Remove legacy agent implementations
- Removed unused and redundant agent files:
  - `src/core/agents/personal-assistant.js`
  - `src/core/agents/simple-master-agent.js`
  - `src/core/agents/simple-personal-assistant.js`

- Refactored `enhanced-personal-assistant.js` to remove inheritance from the legacy `simple-personal-assistant.js`.
- Updated `team-orchestrator.js` to use the correct `master-executive-agent.js`.
- Added missing methods (`getAgentStats`, `forceGenerateSummary`) to `master-executive-agent.js` to ensure compatibility with existing tests.
